Refactoring management classes
Changing all expense, label, and pay to management classes to all
extend from a single management class. This is to enforce a design
pattern and easily give them all the benefits of caching.

A new pay to was added to the DB. “No One” now has index 1 and is the
default for all new expenses. This is a better solution than the
previous solution because now the default comes from the database and
not from the code.